Document new IoMethod trait
